# Use link_set_content_id in Queries.Links