--- a/config/brakeman.ignore
+++ b/config/brakeman.ignore
@@ -18,18 +18,21 @@
       },
       "user_input": "sql",
       "confidence": "Weak",
+      "cwe_id": [
+        89
+      ],
       "note": "The included SQL is a constant string with no user input."
     },
     {
       "warning_type": "SQL Injection",
       "warning_code": 0,
-      "fingerprint": "d0811d2c0f8a29d7aca527231dab7c00fcc62bd08556ba114c0a934618ecc9d7",
+      "fingerprint": "683209c49b28a5b8dcf1f540dbba77d2ef430831a7b3391c7484bd0418bba104",
       "check_name": "SQL",
       "message": "Possible SQL injection",
       "file": "app/queries/links.rb",
-      "line": 211,
+      "line": 210,
       "link": "https://brakemanscanner.org/docs/warning_types/sql_injection/",
-      "code": "Arel.sql(\"\\n        EXISTS(\\n          SELECT nested_links.id\\n          FROM links AS nested_links\\n          INNER JOIN link_sets AS nested_link_sets\\n          ON nested_link_sets.id = nested_links.link_set_id\\n          WHERE #{where}\\n          LIMIT 1\\n        )\\n      \")",
+      "code": "Arel.sql(\"\\n        EXISTS(\\n          SELECT nested_links.id\\n          FROM links AS nested_links\\n          WHERE link_set_content_id IS NOT NULL\\n          AND #{where}\\n          LIMIT 1\\n        )\\n      \")",
       "render_path": null,
       "location": {
         "type": "method",
@@ -38,6 +41,9 @@
       },
       "user_input": "where",
       "confidence": "Medium",
+      "cwe_id": [
+        89
+      ],
       "note": "This SQL is generated using valid content IDs and configuration data, not with arbitrary user input"
     },
     {
@@ -58,6 +64,9 @@
       },
       "user_input": "{ :draft => 0, :published => 1, :unpublished => 1, :superseded => 2 }.slice(*states).map do\n \"WHEN '#{k}' THEN #{v} \"\n end.join",
       "confidence": "Medium",
+      "cwe_id": [
+        89
+      ],
       "note": "The included SQL are constant strings and integers with no user input."
     },
     {
@@ -78,9 +87,12 @@
       },
       "user_input": "values.size",
       "confidence": "Medium",
+      "cwe_id": [
+        89
+      ],
       "note": "The SQL is generated only with valid locales and edition states, not with arbitrary user input"
     }
   ],
-  "updated": "2021-06-10 09:17:56 +0000",
-  "brakeman_version": "5.0.4"
+  "updated": "2025-03-31 11:08:00 +0000",
+  "brakeman_version": "7.0.0"
 }